requirements.txt
wheel
smbus; sys_platform=='Linux'
swig
gpiozero
pyserial
rplcd; sys_platform=='Linux'
requests
Pillow
thermalprinter
numpy
argparse
keyboard; sys_platform=='Windows'